requirements.txt
backcall==0.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11.6
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-guardian==1.5.0
django-import-export==1.2.0
django-reversion==3.0.3
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.12.0
idna==2.8
ipython==7.2.0
ipython-genutils==0.2.0
itypes==1.1.0
jdcal==1.4
jedi==0.13.2
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
mysqlclient==1.4.2
odfpy==1.4.0
openpyxl==2.5.14
parso==0.3.1
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.7
Pygments==2.3.1
PyJWT==1.7.1
pytz==2017.2
PyYAML==3.13
requests==2.21.0
six==1.12.0
tablib==0.12.1
traitlets==4.3.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
